✨ :: [#18] Post Create 기능 구현

+++ Projects/Domain/Sources/Request/Post/CreatePostRequest.swift
@@ -3,10 +3,10 @@ import Foundation
 public struct CreatePostRequest: Codable {
     var title: String
     var location: String
-    var userList: [Int64]
+    var userList: [Int]
     var imageUrl: [String]
 
-    public init(title: String, location: String, userList: [Int64], imageUrl: [String]) {
+    public init(title: String, location: String, userList: [Int], imageUrl: [String]) {
         self.title = title
         self.location = location
         self.userList = userList
